% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kube_install.R
\name{binary_install}
\alias{binary_install}
\title{Install and create binaries for R packages.}
\usage{
binary_install(pkg, lib_path, bin_path)
}
\arguments{
\item{pkg}{character() name of R or Bioconductor package.}

\item{lib_path}{character() path where R package libraries are
stored.}

\item{bin_path}{character() path where R package binaries are
stored.}
}
\value{
`binary_install()` returns invisibly.
}
\description{
Install and create binaries for R packages.
}
\details{
The package given by `pkg` is installed in the given
    library path `lib_path`, and the binaries are created in the
    `bin_path`.
}
\examples{
binary_install(
    pkg = 'AnVIL',
    lib_path = "/host/library",
    bin_path = "/host/binaries"
)

}
