% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_dependency_graph.R
\name{.create_library_paths}
\alias{.create_library_paths}
\title{Create host directories if they don't exist already}
\usage{
.create_library_paths(library_path, binary_path)
}
\description{
Create host directories if they don't exist already
}
\keyword{internal}
