% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_dependency_graph.R
\name{.diff_PACKAGES}
\alias{.diff_PACKAGES}
\title{Compare remote PACKAGES file to current Bioconductor
    PACKAGES file to return list to be updated.}
\usage{
.diff_PACKAGES(bucket_url)
}
\value{
`.diff_PACAKGES(bucket_url)` returns character vector of
    packages to be updated.
}
\description{
Compare remote PACKAGES file to current Bioconductor
    PACKAGES file to return list to be updated.
}
\keyword{internal}
