% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud_sync.R
\name{.gcloud_service_account_auth}
\alias{.gcloud_service_account_auth}
\title{Authenticate with gcloud service account.}
\usage{
.gcloud_service_account_auth(secret)
}
\arguments{
\item{secret}{character(1) path to the location of the secret key
for the service account.}
}
\value{
`.gcloud_auth()` shows a message on successful
    authentication.
}
\description{
Authenticate with gcloud service account.
}
\details{
The function requires a secret service account key file
    which is shared with the kubernetes cluster. The service
    account secret has minimum requirements to run functionality
    for `gsutil mb` and `gsutil rsync`.
}
\keyword{internal}
