% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud_sync.R
\name{.gsutil_make_bucket}
\alias{.gsutil_make_bucket}
\title{Make a bucket on the google cloud}
\usage{
.gsutil_make_bucket(
  bucket,
  uniform_bucket_level_access = FALSE,
  storage_class = "standard",
  location = "us"
)
}
\arguments{
\item{bucket}{character(1) bucket name for the google storage
bucket.}

\item{uniform_bucket_level_access}{"-b" Specifies the uniform
bucket-level access setting. Default is "off"}

\item{storage_class}{"-c" standard: >99.99% in multi-regions and
dual-regions}

\item{location}{"-l" bucket is created in the location US, which is
multi-region}
}
\value{
`.gcloud_make_bucket()` returns invisibly
}
\description{
Make a bucket on the google cloud
}
\keyword{internal}
