% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kube_install.R
\name{.kube_wait}
\alias{.kube_wait}
\title{Wait for worker pods to become active.}
\usage{
.kube_wait(workers)
}
\arguments{
\item{workers}{integer() number of workers in the kubernetes cluster.}
}
\description{
Wait for worker pods to become active.
}
\keyword{internal}
