% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_dependency_graph.R
\name{.packages_to_update}
\alias{.packages_to_update}
\title{Compare binary PACKAGES file to current Bioconductor
    PACKAGES file to return list to be updated.}
\usage{
.packages_to_update(binary_repo = character())
}
\arguments{
\item{binary_repo}{character() vector pointing to binary 
repository which has PACKAGES file.}
}
\value{
`.packages_to_update()` returns character vector of
    packages to be updated.
}
\description{
Compare binary PACKAGES file to current Bioconductor
    PACKAGES file to return list to be updated.
}
\examples{
\dontrun{
repo <- "https://storage.googleapis.com/anvil-rstudio-bioconductor-test/0.99/3.11/"
.packages_to_update(binary_repo = repo)
}

}
\keyword{internal}
