% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_dependency_graph.R
\name{.read_PACKAGES}
\alias{.read_PACKAGES}
\title{Read a PACKAGES file from url}
\usage{
.read_PACKAGES(path)
}
\value{
`.read_PACAKGES(path)` returns the packages and version
    number seperated by an '_'.
}
\description{
Read a PACKAGES file from url
}
\keyword{internal}
