% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud_sync.R
\name{gcloud_binary_sync}
\alias{gcloud_binary_sync}
\title{Sync binaries with Google bucket.}
\usage{
gcloud_binary_sync(bin_path, bucket, secret = "/home/rstudio/key.json")
}
\arguments{
\item{bin_path}{character(1) path to the directory where binaries
are stored.}

\item{bucket}{character(1) bucket name for the google storage
bucket, it must include the FULL path of the CRAN style repo.}

\item{secret}{character(1) path to the location of the secret key
for the service account.}
}
\value{
`gcloud_binary_sync()` return invisibly, but a message is
    shown on screen on successfully tranferred files.
}
\description{
Sync binaries with Google bucket.
}
\details{
Sync binaries created by the `kube_install()` function
    into the Google bucket which is provided. The google bucket
    should be appropriately created in a CRAN style manner with the
    correct public permissions.
}
\examples{
\dontrun{

gcloud_binary_sync(
    bin_path = "/host/binaries",
    bucket = "bioconductor_docker/packages/3.12/bioc/src/contrib/",
    secret = "/home/rstudio/key.json"
)
}

}
\seealso{
`BiocKuberInstall::gcloud_create_cran_bucket()`
}
