% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud_sync.R
\name{gcloud_create_cran_bucket}
\alias{gcloud_create_cran_bucket}
\title{Create a CRAN style google bucket with appropriate ACL}
\usage{
gcloud_create_cran_bucket(
  bucket,
  image_version,
  bioc_version = as.character(BiocManager::version()),
  secret,
  public = TRUE
)
}
\arguments{
\item{bucket}{character(1) bucket name for the google storage
bucket.}

\item{image_version}{characater(1) version number for the
docker image on which binaries are being created.}

\item{bioc_version}{character(1) version number for Bioconductor,
defaults to `BiocManager::version()`.}

\item{secret}{character(1) path to the location of the secret key
for the service account.}

\item{public}{logical(1) if the bucket should be publicly
accessible. Defauly is TRUE, if FALSE the bucket is made
private.}
}
\value{
`gcloud_create_cran_bucket()` returns a character vector of
    the path of the workspace bucket.
}
\description{
Create a CRAN style google bucket with appropriate ACL
}
\details{
Create a CRAN style bucket used to store the package
    binaries built for a specific docker image. The CRAN style
    bucket takes is created using the following format,
    "bucket_name/production_version/bioc_version/src/contrib". All
    buckets created with this function come with Uniform
    Bucket-Level Access and are made public. To create a private
    bucket, toggle the `public` argument to FALSE. The bucket once
    made public is available at the URL provided by google,
    `https://storage.googleapis.com/`.

    An example location for a package would be,
    https://storage.googleapis.com/bucket-name/0.99/3.11/src/contrib/ABAData_1.19.0_R_x86_64-pc-linux-gnu.tar.gz
}
\examples{
\dontrun{
gcloud_create_cran_bucket(
    bucket = "gs://bioconductor-docker-test",
    image_version = "0.99",
    bioc_version = "3.11",
    secret = "/home/mysecret.json",
    public = TRUE
)
}

}
