% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcloud_sync.R
\name{gcloud_sync_to_bucket}
\alias{gcloud_sync_to_bucket}
\title{Sync binaries with Google bucket.}
\usage{
gcloud_sync_to_bucket(src_path, bucket)
}
\arguments{
\item{src_path}{character(1) path to the directory where source
information are stored.}

\item{bucket}{character(1) bucket name for the google storage
bucket, it must include the FULL path of the CRAN style repo.}
}
\value{
`gcloud_sync_to_bucket()` return invisibly, but a message is
    shown on screen on successfully tranferred files.
}
\description{
Sync binaries with Google bucket.
}
\details{
Sync binaries created by the `kube_install()` function
    into the Google bucket which is provided. The google bucket
    should be appropriately created in a CRAN style manner with the
    correct public permissions.
}
\examples{
\dontrun{

gcloud_sync_to_bucket(
    src_path = "/host/binaries",
    bucket = "bioconductor_docker/packages/3.12/bioc/src/contrib/"
)

gcloud_sync_to_bucket(
    src_path = "/host/logs",
    bucket = "bioconductor_docker/packages/3.13/bioc/src/package_logs/"
)
}

}
\seealso{
`BiocKuberInstall::gcloud_create_cran_bucket()`
}
