% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kube_install.R
\name{kube_install_single_package}
\alias{kube_install_single_package}
\title{Install and create binaries for R packages.}
\usage{
kube_install_single_package(pkg, lib_path, bin_path, logs_path)
}
\arguments{
\item{pkg}{character() name of R or Bioconductor package.}

\item{lib_path}{character() path where R package libraries are
stored.}

\item{bin_path}{character() path where R package binaries are
stored.}

\item{logs_path}{character() path where R package binary build logs
are stored.}
}
\value{
`kube_install_single_package()` returns invisibly.
}
\description{
Install and create binaries for R packages.
}
\details{
The package given by `pkg` is installed in the given
    library path `lib_path`, and the binaries are created in the
    `bin_path`.
}
\examples{
\dontrun{
kube_install_single_package(
    pkg = 'AnVIL',
    lib_path = "/host/library",
    bin_path = "/host/binaries"
)
}

}
