% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kube_install.R
\name{kube_wait}
\alias{kube_wait}
\title{Wait for worker pods to become active.}
\usage{
kube_wait(workers = as.integer(1))
}
\arguments{
\item{workers}{integer() number of workers in the kubernetes cluster.}
}
\description{
Wait for kubernetes workers
}
\details{
Wait for the workers to start up. More details on redis
    flags here https://redis.io/commands/client-list.
}
\examples{
\dontrun{
kube_wait(workers = 6L)
}

}
