% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{local_create_cran_repo}
\alias{local_create_cran_repo}
\title{Create a CRAN style local repository}
\usage{
local_create_cran_repo(
  repo,
  bioc_version = as.character(BiocManager::version()),
  volume_mount_path = "/host/"
)
}
\arguments{
\item{repo}{The folder that will contain this repository. Can be set via
\code{BIOCONDUCTOR_BINARY_REPOSITORY} environment variable or option.}

\item{bioc_version}{character(1) version number for Bioconductor,
defaults to `BiocManager::version()`.}
}
\value{
\code{local_create_cran_repo} returns a character vector of the path to
the binary repository.
}
\description{
Create a CRAN style local repository
}
\examples{
\dontrun{
    local_create_cran_repo(repo = "~/dockerhome/.cache/R-crancache")
}
}
