% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_sync.R
\name{local_sync}
\alias{local_sync}
\alias{local_create_cran_bucket}
\title{Create a CRAN style local repository}
\usage{
local_create_cran_bucket(
  folder,
  bioc_version,
  secret = NULL,
  public = TRUE,
  bucket
)
}
\arguments{
\item{bioc_version}{character(1) version number for Bioconductor,
defaults to `BiocManager::version()`.}

\item{secret}{character(1) path to the location of the secret key
for the service account.}

\item{public}{logical(1) if the bucket should be publicly
accessible. Defauly is TRUE, if FALSE the bucket is made
private.}

\item{bucket}{`character(1)` The folder that will contain this repository.
Can be set via `BIOCONDUCTOR_BINARY_REPOSITORY` environment variable or
option.}
}
\value{
`local_create_cran_bucket` returns a character vector of the path to
    the binary repository.
}
\description{
Create a CRAN style local repository
}
\examples{
local_create_cran_bucket(
    folder = "bioconductor_docker",
    bioc_version = "3.21",
    bucket = "/host/"
)
}
