% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_dependency_graph.R
\name{.pkg_dependencies}
\alias{.pkg_dependencies}
\title{Calculate the package dependency graph}
\usage{
.pkg_dependencies(binary_repo = character())
}
\description{
Calculate the package dependency graph
}
\keyword{internal}
