% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kube_install.R
\name{kube_install}
\alias{kube_install}
\title{Install and create binaries for packages parallely using a
kubernetes cluster.}
\usage{
kube_install(
  workers,
  lib_path,
  bin_path,
  deps = .pkg_dependecies(),
  inst = installed.packages()
)
}
\arguments{
\item{workers}{numeric() number of workers in the kubernetes
cluster. It should match the `parallelism` argument in the
k8sredis yaml file.}

\item{lib_path}{character() path where R package libraries are
stored.}

\item{bin_path}{character() path where R package binaries are
stored.}

\item{deps}{package dependecy graph as computed by
`.pkg_dependecies()`.}

\item{inst}{installed packages as computed by
`installed.packages()`.}
}
\description{
Install packages and create binaries using a
    kubernetes cluster for a specific bioconductor docker
    image. The kube_install function can be scaled to a large
    cluster to reduce times even further (in theory). Please note
    that this command will charge your google billing account,
    beware of the charges.
}
\examples{
\donttest{
deps_rds <- readRDS(system.file("extdata", "pkg_dependencies.rds"))
kube_install(
    workers = 6L,
    lib_path = "/host/library",
    bin_path = "/host/binaries",
    deps = deps_rds,
    inst = installed.packages()
)
}

}
