% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kube_install.R
\name{kube_run}
\alias{kube_run}
\title{Run builder on k8s}
\usage{
kube_run(
  bioc_version,
  image_name,
  volume_mount_path = "/host/",
  cloud_id = c("local", "google", "azure"),
  build = c("_software", "_update", "_timings"),
  depth0 = FALSE,
  dry.run = TRUE,
  ultimate_pkg = character(),
  exclude_pkgs = character()
)
}
\arguments{
\item{bioc_version}{character(), bioconductor bioc_version number, e.g 3.12
or 3.13}

\item{image_name}{character(), name of the image for which binaries
are being built}

\item{volume_mount_path}{character(), path to volume mount}

\item{depth0}{logical(1), whether to test only the first layer of packages
with zero dependencies}

\item{dry.run}{logical(1), whether to generate a test run with artificial
artifacts rather than binaries; should be used with `cloud_id = "local"`}

\item{ultimate_pkg}{`character(1)` name of the leaf node package to build
including all its dependencies. This is useful for building a single
package and all its dependencies.}

\item{exclude_pkgs}{character(), list of packages to exclude}

\item{secret}{character(1) path to the location of the secret key
for the service account.}
}
\description{
Run binary installation on k8s cluster
}
\examples{
\dontrun{

kube_run(
    bioc_version = '3.20',
    image_name = 'bioconductor_docker',
    cloud_id = 'local',
    volume_mount_path = '/host/',
    exclude_pkgs = c('canceR')
)
}

}
