% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kube_install.R
\name{kube_run}
\alias{kube_run}
\title{Run builder on k8s}
\usage{
kube_run(
  version,
  image_name,
  workers,
  volume_mount_path = "/host/",
  exclude_pkgs = c("canceR", "flowWorkspace", "gpuMagic", "ChemmineOB")
)
}
\arguments{
\item{version}{character(), bioconductor version number, e.g 3.12
or 3.13}

\item{image_name}{character(), name of the image for which binaries
are being built}

\item{workers}{integer(), number of workers pods in the
k8s cluster}

\item{volume_mount_path}{character(), path to volume mount}

\item{exclude_pkgs}{character(), list of packages to exclude}
}
\description{
Run binary installation on k8s cluster
}
\examples{
\dontrun{

kube_run(version = '3.13',
         image_name = 'bioconductor_docker',
         workers = '10', volume_moun_path = '/host/')

}

}
